+++ scripts/sql_s22_batch.sql
@@ -1,0 +1,89 @@
+BEGIN;
+
+-- Tu'i Tonga (id=44) - 950-1865 Pacific maritime empire
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(44, 'Bott, E. with Tavi (1982). Tongan Society at the Time of Captain Cook''s Visits: Discussions with Her Majesty Queen Salote Tupou. Polynesian Society Memoir 44. Wellington.', 'https://www.thepolynesiansociety.org/publications/', 'academic'),
+(44, 'Gunson, N. (1997). Great Families of Polynesia: Inter-island Links and Marriage Patterns. Journal of Pacific History, 32(2), 139-179. DOI:10.1080/00223349708572834.', 'https://doi.org/10.1080/00223349708572834', 'academic'),
+(44, 'St. Cartmail, K. (1997). The Art of Tonga. University of Hawai''i Press. ISBN 978-0824819729.', 'https://uhpress.hawaii.edu/title/the-art-of-tonga/', 'academic'),
+(44, 'Kirch, P. V. (1984). The Evolution of the Polynesian Chiefdoms. Cambridge University Press. ISBN 978-0521273169.', 'https://www.cambridge.org/9780521273169', 'academic'),
+(44, 'Burley, D. V. (1998). Tongan Archaeology and the Tongan Past, 2850-150 B.P. Journal of World Prehistory, 12(3), 337-392. DOI:10.1023/A:1022322303769.', 'https://doi.org/10.1023/A:1022322303769', 'academic');
+
+-- Croatia medieval (id=433) - Kraljevina Hrvatska 925-1102
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(433, 'Fine, J. V. A. Jr. (1991). The Early Medieval Balkans: A Critical Survey from the Sixth to the Late Twelfth Century. University of Michigan Press. ISBN 978-0472081493.', 'https://www.press.umich.edu/9358', 'academic'),
+(433, 'Curta, F. (2006). Southeastern Europe in the Middle Ages, 500-1250. Cambridge University Press. ISBN 978-0521815390.', 'https://www.cambridge.org/9780521815390', 'academic'),
+(433, 'Budak, N. (1994). Prva stoljeca Hrvatske [First Centuries of Croatia]. Hrvatska sveucilisna naklada. ISBN 978-9531690324.', 'https://hsn.hr', 'academic'),
+(433, 'Margetic, L. (1999). The Croatian state during the era of rulers from the Croatian national dynasty. In: Croatia in the Early Middle Ages: A Cultural Survey. Philip Wilson Publishers. ISBN 978-0856674990.', 'https://www.worldcat.org/title/croatia-in-the-early-middle-ages/oclc/40755055', 'academic'),
+(433, 'Goldstein, I. (1999). Croatia: A History. McGill-Queen''s University Press. ISBN 978-0773520172.', 'https://www.mqup.ca/croatia--a-history-products-9780773520172.php', 'academic');
+
+-- Meissen Margraviate (id=577) - Markgrafschaft Meissen 929-1423
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(577, 'Althoff, G. (1999). Saxony and the Elbe Slavs in the Tenth Century. In: The New Cambridge Medieval History, Vol. 3, ed. T. Reuter. Cambridge University Press. pp. 267-292. ISBN 978-0521364478.', 'https://doi.org/10.1017/CHOL9780521364478', 'academic'),
+(577, 'Bachrach, D. S. (2013). Henry I of Germany''s 929 military campaign in archaeological perspective. Early Medieval Europe, 21(3), 307-337. DOI:10.1111/emed.12026.', 'https://doi.org/10.1111/emed.12026', 'academic'),
+(577, 'Bachrach, D. S. (2020). The Eastern Campaigns of King Henry II of Germany, 1003-17. Journal of Medieval Military History, 18, 1-36. Boydell Press. ISBN 978-1783275564.', 'https://boydellandbrewer.com/9781783275564', 'academic'),
+(577, 'Hardt, M. (2001). Hesse, Elbe, Saale and the Frontiers of the Carolingian Empire. In: W. Pohl, I. Wood, H. Reimitz (eds.), The Transformation of Frontiers: From Late Antiquity to the Carolingians. Brill. pp. 219-232. ISBN 978-9004111158.', 'https://brill.com/view/title/12244', 'academic'),
+(577, 'Reuter, T. (1991). Germany in the Early Middle Ages c. 800-1056. Longman. ISBN 978-0582490345.', 'https://www.routledge.com/9780582490345', 'academic');
+
+-- Iceland Commonwealth (id=650) - Þjóðveldið íslensku 930-1262
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(650, 'Sigurðsson, J. V. (1999). Chieftains and Power in the Icelandic Commonwealth. Odense University Press. ISBN 978-8778382085.', 'https://www.universitypress.dk', 'academic'),
+(650, 'Karlsson, G. (2000). The History of Iceland. University of Minnesota Press. ISBN 978-0816635887.', 'https://www.upress.umn.edu/book-division/books/the-history-of-iceland', 'academic'),
+(650, 'Bagge, S. (2014). Cross and Scepter: The Rise of the Scandinavian Kingdoms from the Vikings to the Reformation. Princeton University Press. ISBN 978-0691161501.', 'https://press.princeton.edu/books/hardcover/9780691161501/cross-and-scepter', 'academic'),
+(650, 'Jakobsson, S. (2021). All the King''s men. The incorporation of Iceland into the Norwegian Realm. Scandinavian Journal of History, 46(5), 571-592. DOI:10.1080/03468755.2021.1903999.', 'https://doi.org/10.1080/03468755.2021.1903999', 'academic'),
+(650, 'Byock, J. (2001). Viking Age Iceland. Penguin. ISBN 978-0140291155.', 'https://www.penguin.co.uk/books/35590/viking-age-iceland/9780140291155.html', 'academic');
+
+-- Mogadishu Sultanate (id=741) - مقديشو 900-1600
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(741, 'Mukhtar, M. H. (2003). Historical Dictionary of Somalia (New ed.). Scarecrow Press. ISBN 978-0810866041.', 'https://rowman.com/ISBN/9780810866041', 'academic'),
+(741, 'Jama, A. (1996). The Origins and Development of Mogadishu AD 1000 to 1850. Studies in African Archaeology 12. Uppsala University. ISBN 978-9150611236.', 'https://www.diva-portal.org/smash/record.jsf?pid=diva2%3A167611', 'academic'),
+(741, 'Cassanelli, L. V. (1982). The Shaping of Somali Society: Reconstructing the History of a Pastoral People, 1600-1900. University of Pennsylvania Press. ISBN 978-0812278323.', 'https://www.upenn.edu/pennpress/book/13888.html', 'academic'),
+(741, 'Lewis, I. M. (2002). A Modern History of the Somali (4th ed.). James Currey. ISBN 978-0852554838.', 'https://www.jamescurrey.com/9780852554838', 'academic'),
+(741, 'Njoku, R. C. (2013). The History of Somalia. ABC-CLIO/Greenwood. ISBN 978-0313378577.', 'https://www.abc-clio.com/products/a3094c/', 'academic');
+
+-- Mixtec (id=792) - Ñuu Tnoo Huahi Andehui 900-1521
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(792, 'Joyce, A. A. (2009). Mixtecs, Zapotecs, and Chatinos: Ancient Peoples of Southern Mexico. Wiley-Blackwell. ISBN 978-0631209782.', 'https://www.wiley.com/en-us/Mixtecs%2C+Zapotecs%2C+and+Chatinos-p-9780631209782', 'academic'),
+(792, 'Terraciano, K. (2001). The Mixtecs of Colonial Oaxaca: Ñudzahui History, Sixteenth Through Eighteenth Centuries. Stanford University Press. ISBN 978-0804751049.', 'https://www.sup.org/books/title/?id=2247', 'academic'),
+(792, 'Jansen, M. & Pérez Jiménez, G. A. (2011). The Mixtec Pictorial Manuscripts: Time, Agency, and Memory in Ancient Mexico. Brill. ISBN 978-9004193581.', 'https://brill.com/view/title/17697', 'academic'),
+(792, 'Spores, R. (1984). The Mixtecs in Ancient and Colonial Times. University of Oklahoma Press. ISBN 978-0806118987.', 'https://www.oupress.com/9780806118987', 'academic'),
+(792, 'Pohl, J. M. D. (1994). The Politics of Symbolism in the Mixtec Codices. Vanderbilt University Publications in Anthropology 46. ISBN 978-0935462395.', 'https://as.vanderbilt.edu/anthropology/publications-mixtec/', 'academic');
+
+-- Mombasa/Mvita (id=824) - 900-1837 Swahili port
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(824, 'Horton, M. & Middleton, J. (2000). The Swahili: The Social Landscape of a Mercantile Society. Blackwell. ISBN 978-0631189190.', 'https://www.wiley.com/en-us/The+Swahili-p-9780631189190', 'academic'),
+(824, 'Pouwels, R. L. (1987). Horn and Crescent: Cultural Change and Traditional Islam on the East African Coast, 800-1900. Cambridge University Press. ISBN 978-0521523097.', 'https://www.cambridge.org/9780521523097', 'academic'),
+(824, 'Meier, P. (2016). Swahili Port Cities: The Architecture of Elsewhere. Indiana University Press. ISBN 978-0253020666.', 'https://iupress.org/9780253020666/swahili-port-cities/', 'academic'),
+(824, 'Berg, F. J. (1968). The Swahili Community of Mombasa, 1500-1900. Journal of African History, 9(1), 35-56. DOI:10.1017/S0021853700009993.', 'https://doi.org/10.1017/S0021853700009993', 'academic'),
+(824, 'Pearson, M. N. (1998). Port Cities and Intruders: The Swahili Coast, India, and Portugal in the Early Modern Era. Johns Hopkins University Press. ISBN 978-0801859090.', 'https://www.press.jhu.edu/books/title/2185/port-cities-and-intruders', 'academic');
+
+-- Pemba (id=827) - 900-1822 Swahili Indian Ocean port
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(827, 'LaViolette, A. & Fleisher, J. B. (2009). The Urban History of a Rural Place: Swahili Archaeology on Pemba Island, Tanzania, 700-1500 AD. International Journal of African Historical Studies, 42(3), 433-455.', 'https://www.jstor.org/stable/40282193', 'academic'),
+(827, 'Fleisher, J. B. (2010). Swahili Synoecism: Rural Settlements and Town Formation on the Central East African Coast, A.D. 750-1500. Journal of Field Archaeology, 35(3), 265-282. DOI:10.1179/009346910X12707320296740.', 'https://doi.org/10.1179/009346910X12707320296740', 'academic'),
+(827, 'Horton, M. & Middleton, J. (2000). The Swahili: The Social Landscape of a Mercantile Society. Blackwell. ISBN 978-0631189190.', 'https://www.wiley.com/en-us/The+Swahili-p-9780631189190', 'academic'),
+(827, 'Pouwels, R. L. (1987). Horn and Crescent: Cultural Change and Traditional Islam on the East African Coast, 800-1900. Cambridge University Press. ISBN 978-0521523097.', 'https://www.cambridge.org/9780521523097', 'academic'),
+(827, 'Brielle, E. S. et al. (2023). Entwined African and Asian genetic roots of medieval peoples of the Swahili coast. Nature, 615, 866-873. DOI:10.1038/s41586-023-05754-w.', 'https://doi.org/10.1038/s41586-023-05754-w', 'academic');
+
+-- Sofala (id=828) - 950-1898 Mozambique Swahili gold port
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(828, 'Newitt, M. D. (1995). A History of Mozambique. Indiana University Press. ISBN 978-0253340504.', 'https://iupress.org/9780253340504/a-history-of-mozambique/', 'academic'),
+(828, 'Freeman-Grenville, G. S. P. (1975). The East African Coast: Select Documents from the First to the Earlier Nineteenth Century (2nd ed.). Oxford University Press / Rex Collings. ISBN 978-0860361299.', 'https://www.worldcat.org/title/east-african-coast/oclc/2493517', 'academic'),
+(828, 'Pearson, M. N. (1998). Port Cities and Intruders: The Swahili Coast, India, and Portugal in the Early Modern Era. Johns Hopkins University Press. ISBN 978-0801859090.', 'https://www.press.jhu.edu/books/title/2185/port-cities-and-intruders', 'academic'),
+(828, 'Pikirayi, I. (2001). The Zimbabwe Culture: Origins and Decline of Southern Zambezian States. AltaMira Press. ISBN 978-0759101111.', 'https://rowman.com/ISBN/9780759101111', 'academic'),
+(828, 'Mudenge, S. I. G. (1988). A Political History of Munhumutapa c. 1400-1902. Zimbabwe Publishing House / James Currey. ISBN 978-0852550878.', 'https://www.jamescurrey.com', 'academic');
+
+-- Ganzhou Uyghur (id=897) - 甘州回鶻 848-1036
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(897, 'Drompp, M. R. (2005). Tang China and the Collapse of the Uighur Empire: A Documentary History. Brill. ISBN 978-9004141292.', 'https://brill.com/view/title/9928', 'academic'),
+(897, 'Millward, J. A. (2007). Eurasian Crossroads: A History of Xinjiang. Columbia University Press. ISBN 978-0231139243.', 'https://cup.columbia.edu/book/eurasian-crossroads/9780231139243', 'academic'),
+(897, 'Golden, P. B. (2011). Central Asia in World History. Oxford University Press. ISBN 978-0195338195.', 'https://global.oup.com/academic/product/central-asia-in-world-history-9780195338195', 'academic'),
+(897, 'Baumer, C. (2012). The History of Central Asia, Volume Two: The Age of the Silk Roads. I.B. Tauris. ISBN 978-1780768328.', 'https://www.bloomsbury.com/uk/history-of-central-asia-the-age-of-the-silk-roads-9781780768328/', 'academic'),
+(897, 'Bosworth, C. E. & Asimov, M. S. (eds.) (2000). History of Civilizations of Central Asia, Vol. IV: The Age of Achievement, A.D. 750 to the End of the Fifteenth Century, Part Two. UNESCO Publishing. ISBN 978-9231036545.', 'https://unesdoc.unesco.org/ark:/48223/pf0000119212', 'academic');
+
+-- Confidence score upgrades
+UPDATE geo_entities SET confidence_score = 0.85 WHERE id IN (433, 650, 741, 792, 897);
+UPDATE geo_entities SET confidence_score = 0.80 WHERE id IN (44, 577, 824, 827, 828);
+
+COMMIT;
+
+SELECT g.id, g.name_original, g.confidence_score, (SELECT count(*) FROM sources s WHERE s.entity_id=g.id) AS n_src FROM geo_entities g WHERE g.id IN (44, 433, 577, 650, 741, 792, 824, 827, 828, 897) ORDER BY g.id;